requirements.txt
prefect==0.15.1
django==3.1.4
djangorestframework==3.12.2
django-webpack-loader==0.7.0
django_cors_headers==3.6.0
django-mailer==2.1
simplejson==3.17.2
requests==2.25.0
